extract static class LicenseProvider for License info supplying
