+++ VirtoCommerce.Platform.Web/Localizations/en.VirtoCommerce.Platform.json
@@ -21,7 +21,7 @@
     },
     "blades": {
       "system-info": {
-        "title": "Platform Information"
+        "title": "Platform information"
       },
       "asset-list": {
         "title": "Asset management",
